backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-app-backend"
version = "0.1.0"
description = "Backend API for Todo App Phase II"
readme = "README.md"
authors = [{name = "Todo App Team", email = "team@todoapp.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "neon>=0.1.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2",
    "alembic>=1.12.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "better-exceptions>=0.3.3",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "httpx>=0.25.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
line-length = 88
target-version = "py39"

[tool.ruff.isort]
known-first-party = ["backend"]
known-third-party = ["fastapi", "sqlmodel", "uvicorn", "pydantic", "sqlalchemy"]
